CDPD-73927. Ozone snapshot data loss fixes (#143)
Description
CDPD-74188. HDDS-11440. Add a lastTransactionInfo field in SnapshotInfo to check for transactions in flight on the snapshot
CDPD-74241. HDDS-11453. OmSnapshotPurge should be in a different ozone manager double buffer batch
CDPD-74611. HDDS-11408. Snapshot rename table entries are propagated incorrectly on snapshot deletes
CDPD-73924. HDDS-11411. Snapshot garbage collection should not run when the keys are moved from a deleted snapshot to the next snapshot in the chain

Co-authored-by: Swaminathan Balachandran <47532440+swamirishi@users.noreply.github.com>